+++ src/ft_memchr.c
@@ -6,16 +6,16 @@
 /*   By: abarchil <abarchil@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/08/04 21:51:59 by abarchil          #+#    #+#             */
-/*   Updated: 2021/12/19 17:05:05 by abarchil         ###   ########.fr       */
+/*   Updated: 2022/01/13 15:09:52 by abarchil         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../includes/minishell.h"
+#include "../minishell.h"
 
-void	*ft_memchr(const void *s, int c, size_t n)
+void	*ft_memchr(const void *s, int c, int n)
 {
 	unsigned char	*str;
-	size_t			count;
+	int				count;
 
 	str = (unsigned char *)s;
 	count = 0;
