+++ src/ft_memset.c
@@ -6,13 +6,13 @@
 /*   By: abarchil <abarchil@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/01 19:29:24 by abarchil          #+#    #+#             */
-/*   Updated: 2021/12/19 16:59:46 by abarchil         ###   ########.fr       */
+/*   Updated: 2022/01/13 15:09:52 by abarchil         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../includes/minishell.h"
+#include "../minishell.h"
 
-void	*ft_memset(void *s, int c, size_t n)
+void	*ft_memset(void *s, int c, int n)
 {
 	unsigned char	*str;
